Grab the dp2 jar from a github releases page
and install it with the `get_url` ansible module

Up until now I put all the install artifacts in the files folder. This
is a simple method and everything is kept in one place. However for
large binaries like this uberjar it admittedly does not make that much
sense to maintain it in git. In fact Github doesn't like this method
and even prevents it for files > 100 MB.

Instead I'm dropping the uberjar on a release page and install it from
there with the help of the `get_url` ansible module.