Reorganize the secrets
having a group named test and production doesn't work (unless you
defined it in the inventory, i.e. in the 'test' or 'production'
files).

So the whole idea of splitting up the secrets into groups doesn't
work. Instead we put the secrets in host specific files. We keep the
normal settings in host_vars/<host>/vars.yml and the (host specific)
secrets in host_vars/<host>/secrets.yml. The global secrets are kept
in group_vars/all.yml